requirements.txt
pandas
torch
pillow
langchain-openai
modelscope